# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/sales_invoices:
    get:
      tags:
        - SalesInvoices
      summary: List SalesInvoices
      description: Get a paginated list of SalesInvoices with optional filtering and sorting
      operationId: listSalesInvoices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesInvoices
      summary: Create SalesInvoice
      description: Create a new SalesInvoice
      operationId: createSalesInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesInvoiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/sales_invoice_items:
    get:
      tags:
        - SalesInvoiceItems
      summary: List SalesInvoiceItems
      description: Get a paginated list of SalesInvoiceItems with optional filtering and sorting
      operationId: listSalesInvoiceItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalesInvoiceItems
      summary: Create SalesInvoiceItem
      description: Create a new SalesInvoiceItem
      operationId: createSalesInvoiceItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalesInvoiceItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesInvoiceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SalesInvoice:
      type: object
      required:
        - id
        - invoice_number
        - company_id
        - customer_id
        - status
        - invoice_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - outstanding_amount
        - receivable_account_id
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
          maxLength: 40
        sales_order_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        outstanding_amount:
          type: string
        receivable_account_id:
          type: string
          format: uuid
        tax_output_account_id:
          nullable: true
          type: string
          format: uuid
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateSalesInvoiceRequest:
      type: object
      required:
        - invoice_number
        - company_id
        - customer_id
        - status
        - invoice_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - outstanding_amount
        - receivable_account_id
        - posting_state
        - metadata
      properties:
        invoice_number:
          type: string
          maxLength: 40
        sales_order_id:
          nullable: true
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        outstanding_amount:
          type: string
        receivable_account_id:
          type: string
          format: uuid
        tax_output_account_id:
          nullable: true
          type: string
          format: uuid
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    SalesInvoiceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesInvoice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SalesInvoiceItem:
      type: object
      required:
        - id
        - invoice_id
        - company_id
        - item_id
        - revenue_account_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        sales_order_item_id:
          nullable: true
          type: string
          format: uuid
        revenue_account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        metadata:
          type: object
...
    CreateSalesInvoiceItemRequest:
      type: object
      required:
        - invoice_id
        - company_id
        - item_id
        - revenue_account_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - metadata
      properties:
        invoice_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        sales_order_item_id:
          nullable: true
          type: string
          format: uuid
        revenue_account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        metadata:
          type: object
...
    SalesInvoiceItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesInvoiceItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []